channels: move node labeler from protokube
Relocates the control-plane node labeler from protokube to a new
channels/pkg/nodelabeler package and renames it to
BootstrapControlPlaneNodeLabels. Protokube still drives the call for
now via the new import path. This is preparation for running channels
as a static pod that owns both addon application and the labels addons
target.

The labeler's tainter.go scratch types are removed; the new package
inlines only the patch struct it needs.
