--- protokube/pkg/protokube/tainter.go
@@ -1,31 +1,0 @@
-/*
-Copyright 2019 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package protokube
-
-type nodePatch struct {
-	Metadata *nodePatchMetadata `json:"metadata,omitempty"`
-	Spec     *nodePatchSpec     `json:"spec,omitempty"`
-}
-
-type nodePatchMetadata struct {
-	Annotations map[string]string `json:"annotations,omitempty"`
-	Labels      map[string]string `json:"labels,omitempty"`
-}
-
-type nodePatchSpec struct {
-	Unschedulable *bool `json:"unschedulable,omitempty"`
-}
